added acquisition device
